feat!: Remove *Tree methods
Instead of having a separate Set and SetTree method for strings and
nested configuration, SetTree was removed and Set now takes an any value
and automatically performs the correct flattening and batching.

The same is true for Show and ShowTree, where ShowTree was removed and
Show now automatically returns either a map or a single value depending
on the context.